Move mixin into its own subpackage
