+++ src/main/java/me/jellysquid/mods/sodium/common/config/SodiumConfig.java
@@ -39,6 +39,7 @@ public class SodiumConfig {
         this.addMixinRule("features.entity.fast_render", true);
         this.addMixinRule("features.entity.smooth_lighting", true);
         this.addMixinRule("features.gui", true);
+        this.addMixinRule("features.gui.fast_loading_screen", true);
         this.addMixinRule("features.gui.font", true);
         this.addMixinRule("features.item", true);
         this.addMixinRule("features.matrix_stack", true);
